yescrypt: move `encoding` module into `params`
The functions are effectively an implementation detail of the params
encoder/decoder at this point, so it's simpler for refactoring to just
have everything in one module.

Ideally these functions could simply be replaced with `Base64ShaCrypt`,
but before that some additional refactoring is probably in order.
